requirements.txt
python-telegram-bot==21.10
openai-whisper
deep-translator
python-dotenv
